1st commit
Playerunknown's Battlegrounds commonly famous as PUBG, a battle royale multiplayer game developed by the PUBG corporation provided some random game stats in the form of a competition in kaggle.com. The same was taken up as a topic in a workshop conducted by Edureka. This is the file which contains all the data analytics functions and programs done in the python programming language. Description of what is done and that of the steps are mentioned in the notebook file itself in the form of markdowns.